\name{survivalEP-package}
\alias{survivalEP-package}
\alias{survivalEP}
\docType{package}
\title{
\packageTitle{survivalEP}
}
\description{
\packageDescription{survivalEP}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{survivalEP}
\packageIndices{survivalEP}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{survivalEP}

Maintainer: \packageMaintainer{survivalEP}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
